- Added trapManager - Added SpikeTrap Component - Added SpikeTrap Prototype - Edited EnemyAI Prefab

+++ Project2/Assets/_Scripts/Enemy/EnemyHealth.cs
@@ -12,24 +12,24 @@ public class EnemyHealth : MonoBehaviour {
 		enemyManager = GameObject.Find ("Game_Manager").GetComponent<EnemyManager> ();
 	}
 
-	public int currentHealth = 100;
-	const int MAX_HEALTH = 100;
+	public float currentHealth = 100.0f;
+	const float MAX_HEALTH = 100.0f;
 
 	//Take some value of damage. (Can also be used to heal as param can be +/-).
-	public void TakeDamage(int damage){
+	public void TakeDamage(float damage){
 		currentHealth -= damage;
 		CheckForDeath ();
 	}
 
 	//Check to see if we're dead.
 	void CheckForDeath(){
-		if(currentHealth <= 0){
+		if(currentHealth <= 0.0f){
 			Die ();
 		}
 	}
 
 	//Adds health to an enemy
-	void Heal(int amount){
+	void Heal(float amount){
 		currentHealth += amount;
 		if(currentHealth > MAX_HEALTH){
 			currentHealth = MAX_HEALTH;
